chore: move to react + ts + vite
